Support reading 64 bit integers
Some sysfs *_input values may be bigger than 32 bits,
for example power values reported in microwatts.  As such,
read an int64_t value out of sysfs instead of just a uint32_t.
The D-Bus property that will hold the value is also the same
type, an int64_t.

Resolves openbmc/openbmc#2686

+++ sysfs.cpp
@@ -254,14 +254,14 @@ HwmonIO::HwmonIO(const std::string& path) : p(path)
 
 }
 
-uint32_t HwmonIO::read(
+int64_t HwmonIO::read(
         const std::string& type,
         const std::string& id,
         const std::string& sensor,
         size_t retries,
         std::chrono::milliseconds delay) const
 {
-    uint32_t val;
+    int64_t val;
     std::ifstream ifs;
     auto fullPath = sysfs::make_sysfs_path(
             p, type, id, sensor);
